% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalphaNR.R
\name{convex_alphaNR}
\alias{convex_alphaNR}
\title{Estimation of the dispersion parameter for Negative Binomial convex non-negative matrix factorization.}
\usage{
convex_alphaNR(data, k = NULL, specific = FALSE)
}
\arguments{
\item{data}{Numeric matrix of counts data.}

\item{k}{Rank to be used for the convex non-negative matrix factorization}

\item{specific}{Logical. If TRUE obs-spcific dispersion is used in the Negative Binomial model.}
}
\value{
Dispersion parameter for convex Negative Binomial NMF. Either vector of length one or vector of length no. of observations if obs-specific dispersion is used.
}
\description{
Likelihood estimation of the dispersion parameter in Negative Binomial using Newton-Raphson.
The dispersion parameter can be either a vector of length one or vector of length no. of observation if obs-specific dispersion is used.
}
\examples{
alpha <- convex_alphaNR()

}
