% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Deviance_fun.R
\name{gkldev}
\alias{gkldev}
\title{Generalized Kullback-Leibler divergence.}
\usage{
gkldev(y, mu)
}
\arguments{
\item{y}{A vector of observations.}

\item{mu}{A vector of estimates.}
}
\value{
The Generalized Kullback-Leibler divergence between a vector of observations y and one of estimates mu.
}
\description{
Generalized Kullback-Leibler divergence.
}
