% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Deviance_fun.R
\name{l1dev}
\alias{l1dev}
\title{L1 divergence.}
\usage{
l1dev(y, mu)
}
\arguments{
\item{y}{A vector of observations.}

\item{mu}{A vector of estimates.}
}
\value{
The L1 divergence between a vector of observations y and one of estimates mu.
}
\description{
L1 divergence.
}
