% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Deviance_fun.R
\name{nbdev}
\alias{nbdev}
\title{Negative Binomial divergence.}
\usage{
nbdev(y, mu, alpha_n)
}
\arguments{
\item{y}{A vector of observations.}

\item{mu}{A vector of estimates.}

\item{alpha_n}{A vector of dispersion parameter (one for each observation).}
}
\value{
The Negative Binomial divergence between a vector of observations y and one of estimates mu.
}
\description{
Negative Binomial divergence.
}
