% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmfgen.R
\name{nmfgen}
\alias{nmfgen}
\title{A general purpose function to perform traditional and convex NMF with different divergence measures.}
\usage{
nmfgen(
  data,
  rank,
  distribution = c("NegativeBinomial", "Tweedie", "Laplace"),
  method = c("traditional", "convex"),
  alpha,
  pwr,
  wmat,
  hmat,
  w1mat,
  w2mat,
  maxiter,
  tolerance,
  initial,
  smallIter
)
}
\arguments{
\item{data}{A matrix of count data.}

\item{rank}{The rank for the factorization.}

\item{distribution}{The underlying model assumption for the factorization. Can assume the following values: "NegativeBinomial", "Tweedie", or "Laplace".}

\item{method}{The chosen method for the factorization. Can be either "traditional" or "convex".}

\item{alpha}{Dispersion parameter for the Negative Binomial matrix factorization.}

\item{pwr}{Power parameter for the Tweedie matrix factorization.}

\item{wmat}{An initialization value for the W matrix for traditional NMF.}

\item{hmat}{An initialization value for the H matrix for traditional NMF.}

\item{w1mat}{An initialization value for the W1 matrix for convex NMF.}

\item{w2mat}{An initialization value for the W2 matrix for convex NMF.}

\item{maxiter}{The maximum number of iterations for the estimation.}

\item{tolerance}{The tolerance value for convergence.}

\item{initial}{The number of initializations to test.}

\item{smallIter}{The number of iterations for which to run each initialization.}
}
\value{
A list with the values for the estimated matrices (W and H for traditional NMF or W1 and W2 for convex NMF) and the final cost value.
}
\description{
A general purpose function to perform traditional and convex NMF with different divergence measures.
}
