% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powertweedie.R
\name{powertweedie}
\alias{powertweedie}
\title{Profile likelihood estimation of power for the Tweedie distribution}
\usage{
powertweedie(data, rank, method = c("traditional", "convex"), power.vector)
}
\arguments{
\item{data}{A matrix of count data.}

\item{rank}{The rank for the factorization.}

\item{method}{The chosen method for the factorization. Can be either "traditional" or "convex".}

\item{power.vector}{A numeric vector with values between 1 and 2 or equal to 0.}
}
\value{
A list with the estimated power parameter of the Tweedie distribution and the full results from the estimation procedure.
}
\description{
Profile likelihood estimation of power for the Tweedie distribution
}
