% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Deviance_fun.R
\name{unitdev}
\alias{unitdev}
\title{Unit deviance from the Tweedie distribution}
\usage{
unitdev(y, mu, pwr)
}
\arguments{
\item{y}{A vector of observations.}

\item{mu}{A vector of estimates.}

\item{pwr}{Power parameter of the Tweedie distribution.}
}
\value{
The unit deviance value from the Tweedie distribution between observations and estimates.
}
\description{
Unit deviance from the Tweedie distribution
}
\examples{
y <- rnorm(20,5)
mu <- rep(5,20)
unitdev(y,mu,0)
}
